DUMP mode now always requires a --region
The DUMP mode without --region allowed to dump the first invocation of
all the extractible regions in a program. To fix issues/51 we need to
globalize variables during dump. Globalization is made much harder
when using the DUMP all because we may be forced to globalize regions
from different compilation units and that causes name collisions.

Because the DUMP all mode was almost never used in practice and made
the memory_dump code very complex, we have decided to remove it.