Desactivate --static in test_pthread for now
Recent change in #152 breaks --static mode in OpenMP because it adds an
OpenMP dependency to the replay library. If gcc is used to build CERE,
then the dependency uses GOMP runtime which is a bit messy, since the
CERE Omp codelets use llvm iomp runtime.

For now I deactivate the --static mode in the test, but the OpenMP
dependency will probably have to be removed from the replay library
at some point.