Separate cerec in 3 languages frontend
We now have cerec for C, cerec++ for C++ and ceref for Fortran (optional).
Source language detection caused issues at link time in the past, especially when
linking C++ : for instance, the C++ standard lib would not always be found because
clang would be called called instead of clang++.

Now, choosing the right compiler is the task of the target application's build system,
instead of CERE's automatic language detection.